move code to bake module
