+++ orca-bakery/orca-bakery.gradle
@@ -33,4 +33,9 @@ dependencies {
 
   testImplementation(project(":orca-test-groovy"))
   testImplementation("com.github.tomakehurst:wiremock:2.15.0")
+  testImplementation("org.junit.jupiter:junit-jupiter-api")
+  testImplementation("org.assertj:assertj-core")
+  testImplementation("org.mockito:mockito-core:2.25.0")
+
+  testRuntimeOnly("org.junit.jupiter:junit-jupiter-api")
 }
